+++ src/interaction/keyboardcontroller.cpp
@@ -37,8 +37,8 @@ namespace interaction {
 
 void KeyboardControllerFixed::keyPressed(KeyAction action, Key key, KeyModifier modifier) {
 	// TODO package in script
-	const double speed = 2.75;
-	const double dt = _handler->deltaTime();
+	const float speed = 2.75;
+	const float dt = static_cast<float>( _handler->deltaTime());
 	if(action == KeyAction::Press|| action == KeyAction::Repeat) {
 		if (key == Key::S) {
 		    glm::vec3 euler(speed * dt, 0.0, 0.0);
@@ -92,7 +92,7 @@ void KeyboardControllerFixed::keyPressed(KeyAction action, Key key, KeyModifier 
 		    _handler->distanceDelta(dist);
 		}
 		if (key == Key::T) {
-			PowerScaledScalar dist(-speed * pow(10, 11) * dt, 0.0);
+			PowerScaledScalar dist(-speed * pow(10.0f, 11.0f) * dt, 0.0f);
 			_handler->distanceDelta(dist);
 		}
 		//if (key == Keys::G) {
@@ -101,11 +101,11 @@ void KeyboardControllerFixed::keyPressed(KeyAction action, Key key, KeyModifier 
 		//	distanceDelta(dist);
 		//}
 		if (key == Key::Y) {
-			PowerScaledScalar dist(-speed * 100.0 * dt, 6.0);
+			PowerScaledScalar dist(-speed * 100.0f * dt, 6.0f);
 			_handler->distanceDelta(dist);
 		}
 		if (key == Key::H) {
-			PowerScaledScalar dist(speed * 100.0 * dt, 6.0);
+			PowerScaledScalar dist(speed * 100.0f * dt, 6.0f);
 			_handler->distanceDelta(dist);
 		}
